pwais/mft-pg/requirements.txt
attrs
click
imageio
imageio-ffmpeg
mlflow==1.14.1
pandas

